Disable version replacement in ldes consumer
If version replacement is enabled, the LDES feed tries to keep a list
of latest product versions in memory before persisting to the store
during first startup. For large feeds this results in an out of memory error without
any data being persisted to the store.

By disabling version replacement, each version is immediately
persisted to the store when fetched from the LDES feed. Since we
already have the ldes-object service in the application stack that 'materializes' the current
version of an IDPC product in the public graph, it doesn't harm to
keep all versions from the LDES feed in another graph.