Disable folding of effective changes on uuid-generation service
If a delete-insert query is executed in which the uuid of a resource
is removed, we end up with a resource without uuid if fold effective
changes is enabled. The triple '<something> a <type>', occurring in
the delete as well as in the insert block will be folded. Hence, there
is no delta the uuid-generation service will act on.